Add files via upload
# Rock-paper-and-scissor-game-code
Hi guys, this is a rock, paper and scissor game which is to be played between two users in which one one is computer itself that give random number itself and the other player can be user(anyone) . Winner of game:      1.Rock + Scissor =Rock.      2. Rock + Paper = Paper         3.Scissor + paper =Scissor